jupyterlab-dataops/requirements/requirements-bigquery.txt
dbt-bigquery==1.0.0
google-cloud-bigquery[bqstorage,pandas]==2.32.0
google-cloud-compute==1.5.1